refactor(utils): move calculatePadding helper to Button component
This helper is only used in the Button component, there's no reason for it to be a global helper

feature/typescript-utils
