+++ src/styles/style-helpers.js
@@ -16,10 +16,6 @@
 import { css } from '@emotion/core';
 import { transparentize } from 'polished';
 
-import { sizes } from './constants';
-
-const { KILO, MEGA, GIGA } = sizes;
-
 /**
  * Shadows
  */
@@ -107,21 +103,3 @@ export const clearfix = css`
     clear: both;
   }
 `;
-
-export const calculatePadding = ({ theme, size: buttonSize }) => (
-  diff = '0px'
-) => {
-  const sizeMap = {
-    /* eslint-disable max-len */
-    [KILO]: `calc(${theme.spacings.bit} - ${diff}) calc(${theme.spacings.mega} - ${diff})`,
-    [MEGA]: `calc(${theme.spacings.byte} - ${diff}) calc(${theme.spacings.giga} - ${diff})`,
-    [GIGA]: `calc(${theme.spacings.kilo} - ${diff}) calc(${theme.spacings.tera} - ${diff})`
-    /* eslint-enable max-len */
-  };
-
-  if (!sizeMap[buttonSize] && buttonSize) {
-    return null;
-  }
-
-  return sizeMap[buttonSize] || sizeMap.mega;
-};
